refactor folder by state
